Add support for etag and modified dates for resources and implement multiple range retrieval.
Implemented an optional etag and modified date for resources for use
with cache headers so agents can detect if the resource is change in
with HEAD requests. The download helper have been refactored to provide
a new head() method for these kind of responses.

Also implement support for multiple range response and fix regresions
caused by these changes before creating proper tests for the new
feature.